pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent"
version = "0.7.4"
description = "Autonomous self-modifying AI agent platform"
requires-python = ">=3.12"
dependencies = [
    "pydantic-ai[anthropic,openai,groq,google]>=1.0.0,<2.0.0",
    "pydantic-settings>=2.7",
    "discord.py>=2.4",
    "aiohttp>=3.10",
    "aiosqlite>=0.20",
    "asyncpg>=0.30",
    "httpx>=0.27",
    "rich>=13.9",
    "structlog>=24.4",
    "python-dotenv>=1.0",
    "gitpython>=3.1",
    "watchfiles>=1.0",
    "typer>=0.13",
    "numpy>=1.26",
    "sqlite-vec>=0.1.6",
    "mcp>=1.3",
    "markdownify>=0.13",
]

[project.scripts]
agent = "agent.main:cli"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=agent --cov-branch --cov-report=term-missing"
markers = [
    "unit: fast isolated tests",
    "integration: tests that use real local resources like temporary SQLite databases",
]

[tool.coverage.run]
branch = true
source = ["agent"]

[tool.coverage.report]
skip_empty = true
show_missing = true
omit = [
    "tests/*",
]
